SyntheticNetworks.jl

Documentation for SyntheticNetworks.jl

SyntheticNetworks.RandomPowerGrid — Type

RandomPowerGrid(num_layers, n, n0, p, q, r, s, u, sampling, α, β, γ, debug)

Our model uses the following heuristic target function for redundancy/cost optimization when adding individual links:

    f(i,j,G) = (d_G(i,j) + 1) ^ r / (dist_spatial(x_i,x_j))

where d_G(i, j) is the length of a shortest path between nodes i, j in network G. Compare Equation (1) in the paper.

Index

SyntheticNetworks.RandomPowerGrid